chore/base-data-and-code-cleanup (#32)
* chore(backend): fix code style and phpstan warnings + add app name in sender for emails

* feat(backend): update base data, allow MaintainWord use case to not delete words

Added a handleDeletion bool parameter on MaintainWordCommand so that an existing Subject or Qualifier is not deleted when editing it and removing a GrammaticalRole.
This is only used in the LoadBaseDataCommand because in this context Subjects and Qualifiers are not created at the same time, resulting in Qualifier deletion before recreation, triggering db errors if existing nicks are related to them.

* chore(backend): code style fix

+++ backend/src/Application/UseCase/MaintainWord.php
@@ -55,12 +55,12 @@ class MaintainWord implements MaintainWordInterface
                     $spec->getQualifierPosition()
                 )
             );
-        } else {
+        } elseif ($maintainWordCommand->isHandleDeletion()) {
             $this->qualifierService->deleteIfExists($word->getId());
         }
         if ($spec->isAsSubject()) {
             $subject = $this->subjectService->createOrUpdate($word);
-        } else {
+        } elseif ($maintainWordCommand->isHandleDeletion()) {
             $this->subjectService->deleteIfExists($word->getId());
         }
